+++ packages/emoji/src/components/Emoji/NativEmojiInlineText.tsx
@@ -1,17 +1,10 @@
 import { toShort } from 'emoji-toolkit';
-import React, { ReactElement, ReactNode } from 'react';
-import { EmojiPluginTheme } from '../../theme';
-
-export interface NativeEmojiInlineTextProps {
-  theme: EmojiPluginTheme;
-  decoratedText: string;
-  children: ReactNode;
-  className?: string;
-}
+import React, { ReactElement } from 'react';
+import { EmojiInlineTextProps } from '../..';
 
 export default function NativeEmojiInlineText({
   decoratedText,
   children,
-}: NativeEmojiInlineTextProps): ReactElement {
+}: EmojiInlineTextProps): ReactElement {
   return <span title={toShort(decoratedText)}>{children}</span>;
 }