+++ packages/emoji/src/components/Emoji/NativeEmojiImage.tsx
@@ -1,16 +1,11 @@
 import React, { ReactElement } from 'react';
-import { EmojiPluginTheme } from '../../theme';
+import { EmojiImageProps } from '../..';
 import convertShortNameToUnicode from '../../utils/convertShortNameToUnicode';
 import emojiList from '../../utils/emojiList';
 
-export interface NativeEmojiImageProps {
-  emoji: string;
-  theme: EmojiPluginTheme;
-}
-
 export default function NativeEmojiImage({
   emoji,
-}: NativeEmojiImageProps): ReactElement {
+}: EmojiImageProps): ReactElement {
   const unicode = emojiList.list[emoji][0];
   const emojiDisplay = convertShortNameToUnicode(unicode);
   return <>{emojiDisplay}</>;