+++ packages/emoji/src/components/EmojiSelect/Popover/index.tsx
@@ -1,10 +1,11 @@
-import React, { Component, ReactElement, WheelEvent } from 'react';
+import React, { Component, ComponentType, ReactElement, WheelEvent } from 'react';
 import PropTypes from 'prop-types';
 import addEmoji from '../../../modifiers/addEmoji';
 import Groups from './Groups';
 import Nav from './Nav';
 import ToneSelect from './ToneSelect';
 import {
+  EmojiImageProps,
   EmojiPluginStore,
   EmojiPluginTheme,
   EmojiSelectGroup,
@@ -19,21 +20,17 @@ interface PopoverProps {
   emojis: EmojiStrategy;
   toneSelectOpenDelay: number;
   isOpen?: boolean;
-  useNativeArt?: boolean;
+  emojiImage: ComponentType<EmojiImageProps>;
 }
 
 export default class Popover extends Component<PopoverProps> {
   static propTypes = {
-    cacheBustParam: PropTypes.string.isRequired,
-    imagePath: PropTypes.string.isRequired,
-    imageType: PropTypes.string.isRequired,
     theme: PropTypes.object.isRequired,
     store: PropTypes.object.isRequired,
     groups: PropTypes.arrayOf(PropTypes.object).isRequired,
     emojis: PropTypes.object.isRequired,
     toneSelectOpenDelay: PropTypes.number.isRequired,
     isOpen: PropTypes.bool,
-    useNativeArt: PropTypes.bool,
   };
 
   activeEmoji: Entry | null = null;
@@ -134,7 +131,7 @@ export default class Popover extends Component<PopoverProps> {
 
   renderToneSelect = (): ReactElement | null => {
     if (this.state.showToneSelect) {
-      const { theme = {} } = this.props;
+      const { theme = {}, emojiImage } = this.props;
 
       const containerBounds = this.container!.getBoundingClientRect();
       const areaBounds = this.groupsElement!.container!.getBoundingClientRect();
@@ -165,6 +162,7 @@ export default class Popover extends Component<PopoverProps> {
           bounds={bounds}
           toneSet={this.toneSet}
           onEmojiSelect={this.onEmojiSelect}
+          emojiImage={emojiImage}
         />
       );
     }
@@ -178,7 +176,7 @@ export default class Popover extends Component<PopoverProps> {
       groups = [],
       emojis,
       isOpen = false,
-      useNativeArt,
+      emojiImage,
     } = this.props;
     const className = isOpen
       ? theme.emojiSelectPopover
@@ -208,7 +206,7 @@ export default class Popover extends Component<PopoverProps> {
           ref={element => {
             this.groupsElement = element;
           }}
-          useNativeArt={useNativeArt}
+          emojiImage={emojiImage}
           isOpen={isOpen}
         />
         <Nav